refactor: e2e
